Updates base field state for the view types.
